refactor: update context imports to use mixins for devices, keybindings, and recording state
