+++ src/components/device-selector.ts
@@ -4,7 +4,7 @@ import { customElement, property, state } from "lit/decorators.js";
 import {
   devicesContext,
   selectedDeviceContext,
-} from "../contexts/dictation-context.js";
+} from "../contexts/mixins/devices-context.js";
 import SelectStyles from "../styles/select.js";
 import { recordingDevicesChangedEvent } from "../utils/events.js";
